+++ src/containers/add-dish-form.js
@@ -1,0 +1,6 @@
+import { reduxForm } from "redux-form";
+import AddDishForm from "../components/add-dish-form";
+
+export default reduxForm({
+  form: "adddish"
+})(AddDishForm);
